send an i32 instead of a string

+++ cli/src/mctl.rs
@@ -182,7 +182,7 @@ async fn main() -> std::result::Result<(), Box<dyn std::error::Error>> {
                 &opt.socket,
                 "publish_nexus",
                 Some(json!({ "uuid": uuid,
-                    "share" : protocol,
+                    "share" : protocol as i32,
                     "key" : key,
                 })),
             )
